+++ src/main/java/br/com/backend/requisitos/bc/RequisitoBC.java
@@ -15,11 +15,14 @@ import br.com.backend.requisitos.dao.RequisitoDAO;
 import br.com.backend.requisitos.dto.interfaces.RequisitoDTOInteface;
 import br.com.backend.requisitos.dto.model.RequisitoDTODetalhadoModel;
 import br.com.backend.requisitos.dto.model.RequisitoDTOModel;
+import br.com.backend.requisitos.entity.Atividade;
 import br.com.backend.requisitos.entity.Integrante;
 import br.com.backend.requisitos.entity.Projeto;
 import br.com.backend.requisitos.entity.Requisito;
 import br.com.backend.requisitos.enums.CategoriaRequisito;
 import br.com.backend.requisitos.enums.ImportanciaRequisito;
+import br.com.backend.requisitos.enums.PerfilIntegranteProjeto;
+import br.com.backend.requisitos.enums.StatusAtividade;
 import br.com.backend.requisitos.utils.Util;
 
 public class RequisitoBC extends AbstractBusiness<Requisito, Integer> {
@@ -78,7 +81,7 @@ public class RequisitoBC extends AbstractBusiness<Requisito, Integer> {
 	public List<RequisitoDTOModel> listarTudo(Integer idUsuario, Integer idProjeto) {
 		try {
 			List<RequisitoDTOModel> requisitosDTO = new ArrayList<RequisitoDTOModel>();
-			for (Requisito r : requisitoDAO.list(idUsuario, idProjeto)) {
+			for (Requisito r : requisitoDAO.list(idProjeto)) {
 				requisitosDTO.add(new RequisitoDTOModel(r.getId(), r.getIdRequisito().toString(), r.getNome(),
 						r.getDescricao(), r.getImportancia().getValue(), r.getFonte(), r.getCategoria().getValue()));
 			}
@@ -96,11 +99,11 @@ public class RequisitoBC extends AbstractBusiness<Requisito, Integer> {
 			if (projeto == null)
 				throw new Exception("Projeto não encontrado");
 
-			Requisito requisito = requisitoDAO.find(idUsuario, idProjeto, idRequisito);
+			Requisito requisito = requisitoDAO.find(idProjeto, idRequisito);
 			if (requisito == null)
 				throw new Exception("Requisito não encontrado");
 
-			Integrante integrante = (Integrante) integranteDAO.find(requisito.getIntegrante().getId());
+			Integrante integrante = integranteDAO.find(requisito.getIntegrante().getId());
 			if (integrante == null)
 				throw new Exception("Usuario não encontrado");
 
@@ -113,34 +116,61 @@ public class RequisitoBC extends AbstractBusiness<Requisito, Integer> {
 		}
 	}
 
+	@Transactional
 	public void alterar(Integer idUsuario, Integer idProjeto, Integer idRequisito, RequisitoDTOInteface r)
 			throws Exception {
 		try {
-			Projeto projeto = (Projeto) projetoDAO.find(idProjeto);
+			Projeto projeto = projetoDAO.find(idUsuario, idProjeto);
 			if (projeto == null)
 				throw new Exception("Projeto não encontrado");
 
-			Requisito requisito = requisitoDAO.find(idUsuario, idProjeto, idRequisito);
+			Requisito requisito = requisitoDAO.find(idProjeto, idRequisito);
 			if (requisito == null)
 				throw new Exception("Requisito não encontrado");
+			
+			Integrante integrante = integranteDAO.findByIdUsuarioAndIdProjeto(idUsuario, idProjeto);
+			
+			if(
+				integrante.getPerfilIntegranteProjeto().equals(PerfilIntegranteProjeto.VISITANTE) ||
+				integrante.getPerfilIntegranteProjeto().equals(PerfilIntegranteProjeto.DESENVOLVEDOR)
+			)
+				throw new Exception("Integrante não tem permissão para alterar requisito");
 
+			requisito.setIdRequisito(r.getIdRequisito());
+			requisito.setNome(r.getNome());
+			requisito.setDescricao(r.getDescricao());
+			requisito.setImportancia(ImportanciaRequisito.valueString(r.getImportancia()));
+			requisito.setFonte(r.getFonte());
+			requisito.setCategoria(CategoriaRequisito.valueString(r.getCategoria()));
 			requisito.setDataAlteracao(Util.currentDate());
 
-			requisitoDAO.mergeHalf(idRequisito, requisito);
+			requisitoDAO.mergeFull(requisito);
 		} catch (Exception e) {
 			throw e;
 		}
 	}
 
+	@Transactional
 	public void remover(Integer idUsuario, Integer idProjeto, Integer idRequisito) throws Exception {
 		try {
-			Projeto projeto = (Projeto) projetoDAO.find(idProjeto);
+			Projeto projeto = (Projeto) projetoDAO.find(idUsuario, idProjeto);
 			if (projeto == null)
 				throw new Exception("Projeto não encontrado");
 
-			Requisito requisito = requisitoDAO.find(idUsuario, idProjeto, idRequisito);
+			Requisito requisito = requisitoDAO.find(idProjeto, idRequisito);
 			if (requisito == null)
 				throw new Exception("Requisito não encontrado");
+			
+			Integrante integrante = integranteDAO.findByIdUsuarioAndIdProjeto(idUsuario, idProjeto);
+			
+			if(
+				integrante.getPerfilIntegranteProjeto().equals(PerfilIntegranteProjeto.VISITANTE) ||
+				integrante.getPerfilIntegranteProjeto().equals(PerfilIntegranteProjeto.DESENVOLVEDOR)
+			)
+				throw new Exception("Integrante não tem permissão para excluir requisito");
+			
+			for(Atividade a : requisito.getAtividades())
+				if(!a.getStatus().equals(StatusAtividade.CONCLUIDO)) throw new Exception("Atividade deste requisito ainda não concluida."); 
 
 			requisitoDAO.remove(idRequisito);
 		} catch (Exception e) {
